- dont add /dev/shm as tmpfs to fstab. This fixes the network   setup done by /etc/init.d/network
git-svn-id: https://svn.berlios.de/svnroot/repos/kiwi/kiwi-head@796 92bc7f66-5d17-0410-ac0d-e081219a1bb3

+++ modules/KIWIRoot.pm
@@ -450,10 +450,23 @@ sub setup {
 	#----------------------------------------
 	if (-d "$imageDesc/root") {
 		$kiwi -> info ("Copying user defined files to image tree");
-		qx ( mkdir $root/tmproot );
-		qx ( cp -LR --remove-destination $imageDesc/root/* $root/tmproot 2>&1 );
+		mkdir $root."/tmproot";
+		my $copy = "cp -LR --remove-destination";
+		my $data = qx ( $copy $imageDesc/root/* $root/tmproot 2>&1 );
+		my $code = $? >> 8;
+		if ($code != 0) {
+			$kiwi -> failed ();
+			$kiwi -> info   ($data);
+			return undef;
+		}
 		qx ( find $root/tmproot -type d | grep .svn\$ | xargs rm -rf 2>&1 );
-		qx ( cp -LR --remove-destination $root/tmproot/* $root );
+		$data = qx ( $copy $root/tmproot/* $root );
+		$code = $? >> 8;
+		if ($code != 0) {
+			$kiwi -> failed ();
+			$kiwi -> info   ($data);
+			return undef;
+		}
 		qx ( rm -rf $root/tmproot );
 		$kiwi -> done();
 	}
